pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "cece-tools"
version = "0.1.0"
description = "Tools and scripts for the CECE (Atmospheric Chemistry Emissions Stacker) component."
requires-python = ">=3.8"
dependencies = [
    "pyyaml",
]

[project.optional-dependencies]
test = [
    "pytest",
]
